requirements.txt
torch
torchvision
pillow
pygetwindow
